Factor the construction of generic test case views into a reusable category method.
The popover and terminal tests also use generic views but cannot leverage the
category method because they require a reference to the label that's part of the
generic view. Nevertheless, their layout of the label has been updated to fix a
clipping issue discovered in the development of the category method.